preparatory refactoring of `AbstractAnalysis`
