+++ sonar-plugin/sonar-javascript-plugin/src/main/java/org/sonar/plugins/javascript/analysis/AnalysisWithProgram.java
@@ -26,18 +26,13 @@ import java.util.Deque;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import javax.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.scanner.ScannerSide;
-import org.sonar.plugins.javascript.CancellationException;
 import org.sonar.plugins.javascript.JavaScriptPlugin;
-import org.sonar.plugins.javascript.analysis.cache.CacheAnalysis;
-import org.sonar.plugins.javascript.analysis.cache.CacheStrategies;
 import org.sonar.plugins.javascript.bridge.AnalysisWarningsWrapper;
 import org.sonar.plugins.javascript.bridge.BridgeServer;
-import org.sonar.plugins.javascript.bridge.BridgeServer.JsAnalysisRequest;
 import org.sonar.plugins.javascript.bridge.BridgeServer.TsProgram;
 import org.sonar.plugins.javascript.bridge.BridgeServer.TsProgramRequest;
 import org.sonar.plugins.javascript.utils.ProgressReport;
@@ -107,7 +102,7 @@ public class AnalysisWithProgram extends AbstractAnalysis {
         );
         for (var f : skippedFiles) {
           LOG.debug("File not part of any tsconfig.json: {}", f);
-          analyze(f, null);
+          analyzeFile(f, null, null);
         }
       }
       success = true;
@@ -139,7 +134,7 @@ public class AnalysisWithProgram extends AbstractAnalysis {
         continue;
       }
       if (analyzedFiles.add(inputFile)) {
-        analyze(inputFile, program);
+        analyzeFile(inputFile, null, program);
         counter++;
       } else {
         LOG.debug(
@@ -152,54 +147,4 @@ public class AnalysisWithProgram extends AbstractAnalysis {
     LOG.info("Analyzed {} file(s) with current program", counter);
   }
 
-  private void analyze(InputFile file, @Nullable TsProgram tsProgram) throws IOException {
-    if (context.isCancelled()) {
-      throw new CancellationException(
-        "Analysis interrupted because the SensorContext is in cancelled state"
-      );
-    }
-    var cacheStrategy = CacheStrategies.getStrategyFor(context, file);
-    if (cacheStrategy.isAnalysisRequired()) {
-      try {
-        LOG.debug("Analyzing file: {}", file.uri());
-        progressReport.nextFile(file.toString());
-        var fileContent = contextUtils.shouldSendFileContent(file) ? file.contents() : null;
-        var request = getJsAnalysisRequest(file, tsProgram, fileContent);
-
-        var response = isJavaScript(file)
-          ? bridgeServer.analyzeJavaScript(request)
-          : bridgeServer.analyzeTypeScript(request);
-
-        analysisProcessor.processResponse(context, checks, file, response);
-        cacheStrategy.writeAnalysisToCache(
-          CacheAnalysis.fromResponse(response.ucfgPaths(), response.cpdTokens()),
-          file
-        );
-      } catch (IOException e) {
-        LOG.error("Failed to get response while analyzing " + file.uri(), e);
-        throw e;
-      }
-    } else {
-      LOG.debug("Processing cache analysis of file: {}", file.uri());
-      var cacheAnalysis = cacheStrategy.readAnalysisFromCache();
-      analysisProcessor.processCacheAnalysis(context, file, cacheAnalysis);
-    }
-  }
-
-  private JsAnalysisRequest getJsAnalysisRequest(
-    InputFile file,
-    @Nullable TsProgram tsProgram,
-    @Nullable String fileContent
-  ) {
-    return new JsAnalysisRequest(
-      file.absolutePath(),
-      file.type().toString(),
-      inputFileLanguage(file),
-      fileContent,
-      contextUtils.ignoreHeaderComments(),
-      null,
-      tsProgram != null ? tsProgram.programId() : null,
-      analysisMode.getLinterIdFor(file)
-    );
-  }
 }
